telegram-bot/requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dotenv==1.0.0
schedule==1.2.0
Pillow==10.1.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1 